Make From to return normalized value
